Safari-609.1.11.1/LayoutTests/accessibility/mac/aria-liveregions-notifications-always-sent-expected.txt

Test test1
test2
This tests that ARIA live region notifications are sent even when an assistive technology is not accessing them directly on each update.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification1 is true
PASS addedNotification2 is true
PASS successfullyParsed is true

TEST COMPLETE
